Seperate popUp from script.js in a module
