+++ script.js
@@ -1,4 +1,5 @@
 import projects from "../modules/projectsInfo.js";
+import addElement from "./modules/popUp.js";
 // First, from where our work section will be located?
 // We need a id for reference our JS section.
 const works = document.querySelector('.works');
@@ -81,144 +82,6 @@ const btn3 = document.querySelector('.openWindow2');
 const btn4 = document.querySelector('.openWindow3');
 const mainContainer = document.querySelector('.mainContainer');
 
-// Create a div to open the window into the div.
-function addElement() {
-  const modal = document.createElement('div');
-  const divHeader = document.createElement('div');
-  const header = document.createElement('h1');
-  const closeBtn = document.createElement('button');
-  const img = document.createElement('img');
-  const textDescription = document.createElement('p');
-  const technologiesContainer = document.createElement('div');
-  const verticalLine1 = document.createElement('div');
-  const tech1 = document.createElement('p');
-  const verticalLine2 = document.createElement('div');
-  const tech2 = document.createElement('p');
-  const verticalLine3 = document.createElement('div');
-  const tech3 = document.createElement('p');
-  const verticalLine4 = document.createElement('div');
-  const newBtn = document.createElement('button');
-  const newBtnAnchor = document.createElement('a');
-  const newBtn2 = document.createElement('button');
-  const divBtns = document.createElement('div');
-  const imgBtn1 = document.createElement('img');
-  const imgBtn2 = document.createElement('img');
-  const tech4 = document.createElement('p');
-  const verticalLine5 = document.createElement('div');
-
-  modal.className = 'divClass';
-  modal.classList.add('divClassClose');
-  document.body.appendChild(modal);
-
-  // Header and Close Btn
-  divHeader.className = 'divHeader';
-  modal.appendChild(divHeader);
-
-  header.className = 'h1Class';
-  header.textContent = `${projects[0].name}`;
-  divHeader.appendChild(header);
-
-  closeBtn.classList.add('divBtnClass');
-  closeBtn.textContent = 'X';
-  divHeader.appendChild(closeBtn);
-
-  // Image
-  img.className = 'imgClass';
-  img.src = `${projects[0].image}`;
-
-  if (window.matchMedia('(min-width: 1000px)').matches) {
-    img.src = './BG/Snapshoot-Portfolio-desktop.png';
-  }
-
-  img.alt = 'snapshoot-portfolio';
-  modal.appendChild(img);
-
-  // Description
-  textDescription.className = 'pClass';
-
-  if (window.matchMedia('(min-width: 1000px)').matches) {
-    textDescription.textContent = `${projects[0].descriptionDesktopPopUp}`;
-  }
-
-  if (window.matchMedia('(max-width: 999px)').matches) {
-    textDescription.textContent = `${projects[0].descriptionMobilePopup}`;
-  }
-
-  modal.appendChild(textDescription);
-
-  // Technologies
-  technologiesContainer.classList.add('vertical-lines-modal');
-  modal.appendChild(technologiesContainer);
-
-  verticalLine1.classList.add('vertical-line-modal');
-  technologiesContainer.appendChild(verticalLine1);
-
-  verticalLine2.classList.add('vertical-line-modal');
-  technologiesContainer.appendChild(verticalLine2);
-
-  verticalLine3.classList.add('vertical-line-modal');
-  technologiesContainer.appendChild(verticalLine3);
-
-  verticalLine4.classList.add('vertical-line-modal');
-  technologiesContainer.appendChild(verticalLine4);
-
-  tech1.classList.add('techs');
-  tech2.classList.add('techs');
-  tech3.classList.add('techs');
-
-  technologiesContainer.appendChild(tech1);
-  technologiesContainer.appendChild(verticalLine2);
-  technologiesContainer.appendChild(tech2);
-  technologiesContainer.appendChild(verticalLine3);
-  technologiesContainer.appendChild(tech3);
-  technologiesContainer.appendChild(verticalLine4);
-
-  tech1.textContent = `${projects[0].technologies[0]}`;
-  tech2.textContent = `${projects[0].technologies[1]}`;
-  tech3.textContent = `${projects[0].technologies[2]}`;
-
-  if (window.matchMedia('(min-width: 1000px)').matches) {
-    tech4.classList.add('techs');
-    verticalLine5.classList.add('vertical-line-modal');
-    technologiesContainer.appendChild(tech4);
-    technologiesContainer.appendChild(verticalLine5);
-    tech4.textContent = 'Github';
-  }
-
-  // Buttons
-  divBtns.className = 'divBtns';
-  modal.appendChild(divBtns);
-
-  newBtn.className = 'newBtn';
-  newBtn.textContent = 'See Live';
-  newBtn.appendChild(newBtnAnchor);
-  divBtns.appendChild(newBtn);
-
-  newBtn2.className = 'newBtn2';
-  newBtn2.textContent = 'See Source';
-  divBtns.appendChild(newBtn2);
-
-  imgBtn1.className = 'imgClassBtn';
-  imgBtn1.src = './icons/Icon-btn.png';
-  imgBtn1.alt = 'github-icon';
-  newBtn.appendChild(imgBtn1);
-
-  imgBtn2.className = 'imgClassBtn';
-  imgBtn2.src = './icons/github.png';
-  imgBtn2.alt = 'github-icon';
-  newBtn2.appendChild(imgBtn2);
-
-  // Hide an element
-  const hide = (() => {
-    modal.style.display = 'none';
-  });
-
-  closeBtn.addEventListener('click', () => {
-    mainContainer.style.filter = 'blur(0px)';
-    hide();
-  });
-}
-
 // After clicking in the button the window popup to show your project deatails.
 btn.addEventListener('click', () => {
   addElement();
@@ -239,75 +102,3 @@ btn4.addEventListener('click', () => {
   addElement();
   mainContainer.style.filter = 'blur(20px)';
 });
-
-/* Client Validation */
-const form = document.getElementById('signup');
-// let form = document.getElementById('subscribe');
-const { email } = form.elements;
-// const emailAddress = email.value;
-
-const requiredFields = [
-  { input: email, message: 'Email is required' },
-];
-
-function error(input, message) {
-  input.className = 'error';
-  // show the error message
-  const error = document.querySelector('.labelformat');
-  error.innerText = message;
-  return false;
-}
-
-function success(input) {
-  input.className = 'success';
-  // hide the error message
-  const error = input.previousElementSibling;
-  error.innerText = '';
-  return true;
-}
-
-function validateEmail(input) {
-  const re = /^[a-z0-9.!#$%&'*+=?^_`{|}~-]+@[a-z0-9-]+(?:\.[a-z90-]+)*$/;
-
-  return re.test(input.value.trim())
-    ? success(input)
-    : error(input, 'Invalid email format. Use Lowercase');
-}
-
-// checks if an input element has a non-empty
-function requireValue(input, message) {
-  return input.value.trim() === ''
-    ? error(input, message)
-    : success(input);
-}
-
-form.addEventListener('submit', (event) => {
-  // check required fields
-  let valid = true;
-  requiredFields.forEach((input) => {
-    valid = requireValue(input.input, input.message);
-  });
-  // validate email
-  if (valid) {
-    valid = validateEmail(email);
-  }
-  // stop submitting the form if the data is invalid
-  if (!valid) {
-    event.preventDefault();
-  }
-});
-
-/* Local Storage */
-const localStorageName = document.getElementById('name');
-const localStorageEmail = document.getElementById('email');
-const localStoragebtnInsert = document.getElementById('btnInsert');
-
-localStoragebtnInsert.onclick = () => {
-  const key = localStorageName.value;
-  const { value } = localStorageEmail;
-
-  if (key && value) {
-    localStorage.setItem(key, value);
-    window.location.reload();
-  }
-};
